+++ example/lib/main.dart
@@ -1,12 +1,16 @@
 import 'package:logging/logging.dart';
 import 'package:logging_appenders/logging_appenders.dart';
 
-final _logger = Logger('main');
+final _logger = Logger('example');
 
 void main() {
   Logger.root.level = Level.ALL;
-  final appender = PrintAppender()..attachToLogger(Logger.root);
+  final appender = PrintAppender(formatter: const ColorFormatter())
+    ..attachToLogger(Logger.root);
   _logger.fine('Lorem ipsum');
+  _logger.info('An important info message');
+  _logger.severe('This is bad.');
+  _logger.shout('This is just impolite');
 
   // optionally dispose of the appender.
   appender.dispose();