+++ test/test_utils.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:logging/logging.dart';
+
 class TestUtils {
   static T overridePrint<T>(List<String> log, T Function() testFn) {
     final spec = ZoneSpecification(print: (_, __, ___, String msg) {
@@ -8,4 +10,7 @@ class TestUtils {
     });
     return Zone.current.fork(specification: spec).run(testFn);
   }
+
+  static LogRecord logRecord(String message, {Level level = Level.FINE}) =>
+      LogRecord(level, message, 'test_logger');
 }